refactor : 불필요한 인터페이스 제거 및 수동 로또 티켓 생성을 위한 클래스 생성
