--- src/main/java/lotto/domain/LottoMachine.java
@@ -1,7 +1,0 @@
-package lotto.domain;
-
-public interface LottoMachine {
-    LottoTickets createTickets(int numberOfTickets);
-
-    LottoTickets createTicketsByMoney(int money);
-}
