--- src/main/java/lotto/domain/LottoTicketFactory.java
@@ -1,6 +1,0 @@
-package lotto.domain;
-
-public interface LottoTicketFactory {
-    LottoTicket createLottoTicket();
-}
-
